// client/src/app/api/models/participation-list-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ConstructorListDto } from '../models/constructor-list-dto';
import { DriverParticipationDto } from '../models/driver-participation-dto';
export interface ParticipationListDto {
  constructors?: Array<ConstructorListDto> | null;
  constructorsChampId?: string;
  drivers?: Array<DriverParticipationDto> | null;
  driversChampId?: string;
}